etl_project/requirements.txt
elasticsearch==8.6
psycopg2-binary
python-dotenv==0.21
pydantic
backoff
# backoff==1.11
backoff-utils
redis==4.5